algorithm/requirements-dev.txt
pytest>=5.3.2
pytest-coverage>=0.0
requests-mock>=1.7.0
pytest-datadir>=1.3.1
